Move whiteboard socket handlers to its own router file
